+++ scu-gr/src/fiskaltrust.Middleware.SCU.GR.MyData/Models/Overrides/AddressOverride.cs
@@ -1,0 +1,18 @@
+﻿using System.Text.Json.Serialization;
+
+namespace fiskaltrust.Middleware.SCU.GR.MyData;
+
+public class AddressOverride
+{
+    [JsonPropertyName("street")]
+    public string? Street { get; set; }
+
+    [JsonPropertyName("number")]
+    public string? Number { get; set; }
+
+    [JsonPropertyName("postalCode")]
+    public string? PostalCode { get; set; }
+
+    [JsonPropertyName("city")]
+    public string? City { get; set; }
+}
