+++ scu-gr/src/fiskaltrust.Middleware.SCU.GR.MyData/Models/Overrides/ChargeItemMyDataOverride.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json.Serialization;
+
+namespace fiskaltrust.Middleware.SCU.GR.MyData;
+
+public class ChargeItemMyDataOverride
+{
+    [JsonPropertyName("invoiceDetails")]
+    public InvoiceDetailOverride? InvoiceDetails { get; set; }
+}
