+++ scu-gr/src/fiskaltrust.Middleware.SCU.GR.MyData/Models/Overrides/InvoiceHeaderOverride.cs
@@ -1,0 +1,62 @@
+using System;
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace fiskaltrust.Middleware.SCU.GR.MyData;
+
+public class InvoiceHeaderOverride
+{
+    [JsonPropertyName("invoiceType")]
+    public string? InvoiceType { get; set; }
+
+    [JsonPropertyName("vatPaymentSuspension")]
+    public bool? VatPaymentSuspension { get; set; }
+
+    [JsonPropertyName("selfPricing")]
+    public bool? SelfPricing { get; set; }
+
+    [JsonPropertyName("dispatchDate")]
+    public DateTime? DispatchDate { get; set; }
+
+    [JsonPropertyName("dispatchTime")]
+    public DateTime? DispatchTime { get; set; }
+
+    [JsonPropertyName("vehicleNumber")]
+    public string? VehicleNumber { get; set; }
+
+    [JsonPropertyName("movePurpose")]
+    public int? MovePurpose { get; set; }
+
+    [JsonPropertyName("fuelInvoice")]
+    public bool? FuelInvoice { get; set; }
+
+    [JsonPropertyName("specialInvoiceCategory")]
+    public int? SpecialInvoiceCategory { get; set; }
+
+    [JsonPropertyName("invoiceVariationType")]
+    public int? InvoiceVariationType { get; set; }
+
+    [JsonPropertyName("otherDeliveryNoteHeader")]
+    public OtherDeliveryNoteHeaderOverride? OtherDeliveryNoteHeader { get; set; }
+
+    [JsonPropertyName("otherMovePurposeTitle")]
+    public string? OtherMovePurposeTitle { get; set; }
+
+    [JsonPropertyName("exchangeRate")]
+    public decimal? ExchangeRate { get; set; }
+
+    [JsonPropertyName("thirdPartyCollection")]
+    public bool? ThirdPartyCollection { get; set; }
+
+    [JsonPropertyName("totalCancelDeliveryOrders")]
+    public bool? TotalCancelDeliveryOrders { get; set; }
+
+    [JsonPropertyName("reverseDeliveryNote")]
+    public bool? ReverseDeliveryNote { get; set; }
+
+    [JsonPropertyName("reverseDeliveryNotePurpose")]
+    public int? ReverseDeliveryNotePurpose { get; set; }
+
+    [JsonPropertyName("otherCorrelatedEntities")]
+    public List<EntityOverride>? OtherCorrelatedEntities { get; set; }
+}
