+++ scu-gr/src/fiskaltrust.Middleware.SCU.GR.MyData/Models/Overrides/InvoiceOverride.cs
@@ -1,0 +1,15 @@
+using System.Text.Json.Serialization;
+
+namespace fiskaltrust.Middleware.SCU.GR.MyData;
+
+public class InvoiceOverride
+{
+    [JsonPropertyName("invoiceHeader")]
+    public InvoiceHeaderOverride? InvoiceHeader { get; set; }
+
+    [JsonPropertyName("counterpart")]
+    public PartyUnmappedFieldsOverride? Counterpart { get; set; }
+
+    [JsonPropertyName("issuer")]
+    public PartyUnmappedFieldsOverride? Issuer { get; set; }
+}
