+++ scu-gr/src/fiskaltrust.Middleware.SCU.GR.MyData/Models/Overrides/ReceiptRequestMyDataOverride.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Serialization;
+
+namespace fiskaltrust.Middleware.SCU.GR.MyData;
+
+public class ReceiptRequestMyDataOverride
+{
+    [JsonPropertyName("invoice")]
+    public InvoiceOverride? Invoice { get; set; }
+}
