+++ scu-gr/src/fiskaltrust.Middleware.SCU.GR.MyData/MyDataSCU.cs
@@ -11,379 +11,12 @@ using fiskaltrust.ifPOS.v2;
 using fiskaltrust.ifPOS.v2.Cases;
 using fiskaltrust.storage.V0.MasterData;
 using fiskaltrust.Middleware.SCU.GR.MyData.Models;
-using System.Text.Json.Serialization;
 using System.Security.Cryptography;
 using fiskaltrust.ifPOS.v2.gr;
 using fiskaltrust.Middleware.SCU.GR.Abstraction;
 using fiskaltrust.Middleware.SCU.GR.MyData.Helpers;
 
 namespace fiskaltrust.Middleware.SCU.GR.MyData;
-
-public class ftReceiptCaseDataPayload
-{
-    [JsonPropertyName("GR")]
-    public ftReceiptCaseDataGreekPayload? GR { get; set; }
-}
-
-public class ftReceiptCaseDataGreekPayload
-{
-    public string? MerchantVATID { get; set; }
-    public string? Series { get; set; }
-    public long? AA { get; set; }
-    public string? HashAlg { get; set; }
-    public string? HashPayload { get; set; }
-    
-    /// <summary>
-    /// MyData override configuration allowing direct control of invoice properties
-    /// </summary>
-    [JsonPropertyName("mydataoverride")]
-    public MyDataOverride? MyDataOverride { get; set; }
-}
-
-public class MyDataOverride
-{
-    /// <summary>
-    /// Invoice-level overrides
-    /// </summary>
-    [JsonPropertyName("invoice")]
-    public InvoiceOverride? Invoice { get; set; }
-
-    /// <summary>
-    /// Invoice details (line-item) overrides (applied via ftChargeItemCaseData)
-    /// </summary>
-    [JsonPropertyName("invoicedetails")]
-    public InvoiceDetailsOverride? InvoiceDetails { get; set; }
-}
-
-public class InvoiceDetailsOverride
-{
-    /// <summary>
-    /// Right to deduct VAT for this line.
-    /// In Greek VAT context, "δικαίωμα έκπτωσης" = whether the VAT of that specific line is deductible (input VAT deduction) for the buyer/recipient.
-    /// </summary>
-    [JsonPropertyName("discountOption")]
-    public bool? DiscountOption { get; set; }
-}
-
-/// <summary>
-/// Payload wrapper for ftChargeItemCaseData - mirrors ftReceiptCaseDataPayload but for charge items.
-/// </summary>
-public class ftChargeItemCaseDataPayload
-{
-    [JsonPropertyName("GR")]
-    public ftChargeItemCaseDataGreekPayload? GR { get; set; }
-}
-
-public class ftChargeItemCaseDataGreekPayload
-{
-    [JsonPropertyName("mydataoverride")]
-    public ChargeItemMyDataOverride? MyDataOverride { get; set; }
-}
-
-public class InvoiceOverride
-{
-    /// <summary>
-    /// Invoice header overrides (maps to InvoiceHeaderType in the MyData XML)
-    /// </summary>
-    [JsonPropertyName("invoiceHeader")]
-    public InvoiceHeaderOverride? InvoiceHeader { get; set; }
-
-    /// <summary>
-    /// Counterpart (buyer) overrides for fields not mapped from cbCustomer.
-    /// Only documentIdNo, supplyAccountNo, and countryDocumentId can be set.
-    /// </summary>
-    [JsonPropertyName("counterpart")]
-    public PartyUnmappedFieldsOverride? Counterpart { get; set; }
-
-    /// <summary>
-    /// Issuer overrides for fields not mapped from master data.
-    /// Only documentIdNo, supplyAccountNo, and countryDocumentId can be set.
-    /// </summary>
-    [JsonPropertyName("issuer")]
-    public PartyUnmappedFieldsOverride? Issuer { get; set; }
-}
-
-public class InvoiceHeaderOverride
-{
-    /// <summary>
-    /// Invoice type override. Only allowed values: 3.1, 3.2, 6.1, 6.2, 8.1, 8.2, 9.3
-    /// Can only be set when the automatically determined invoice type would result in an error.
-    /// </summary>
-    [JsonPropertyName("invoiceType")]
-    public string? InvoiceType { get; set; }
-
-    /// <summary>
-    /// VAT payment suspension indicator
-    /// </summary>
-    [JsonPropertyName("vatPaymentSuspension")]
-    public bool? VatPaymentSuspension { get; set; }
-
-    /// <summary>
-    /// Self-pricing indicator
-    /// </summary>
-    [JsonPropertyName("selfPricing")]
-    public bool? SelfPricing { get; set; }
-
-    /// <summary>
-    /// Dispatch date (format: yyyy-MM-dd)
-    /// </summary>
-    [JsonPropertyName("dispatchDate")]
-    public DateTime? DispatchDate { get; set; }
-
-    /// <summary>
-    /// Dispatch time (format: HH:mm:ss)
-    /// </summary>
-    [JsonPropertyName("dispatchTime")]
-    public DateTime? DispatchTime { get; set; }
-
-    /// <summary>
-    /// Vehicle number for transport documents
-    /// </summary>
-    [JsonPropertyName("vehicleNumber")]
-    public string? VehicleNumber { get; set; }
-
-    /// <summary>
-    /// Move purpose code
-    /// </summary>
-    [JsonPropertyName("movePurpose")]
-    public int? MovePurpose { get; set; }
-
-    /// <summary>
-    /// Fuel invoice indicator
-    /// </summary>
-    [JsonPropertyName("fuelInvoice")]
-    public bool? FuelInvoice { get; set; }
-
-    /// <summary>
-    /// Special invoice category code
-    /// </summary>
-    [JsonPropertyName("specialInvoiceCategory")]
-    public int? SpecialInvoiceCategory { get; set; }
-
-    /// <summary>
-    /// Invoice variation type code
-    /// </summary>
-    [JsonPropertyName("invoiceVariationType")]
-    public int? InvoiceVariationType { get; set; }
-
-    /// <summary>
-    /// Other delivery note header information
-    /// </summary>
-    [JsonPropertyName("otherDeliveryNoteHeader")]
-    public OtherDeliveryNoteHeaderOverride? OtherDeliveryNoteHeader { get; set; }
-
-    /// <summary>
-    /// Other move purpose title (free text)
-    /// </summary>
-    [JsonPropertyName("otherMovePurposeTitle")]
-    public string? OtherMovePurposeTitle { get; set; }
-
-    /// <summary>
-    /// Exchange rate when currency is not EUR
-    /// </summary>
-    [JsonPropertyName("exchangeRate")]
-    public decimal? ExchangeRate { get; set; }
-
-    /// <summary>
-    /// Third party collection indicator
-    /// </summary>
-    [JsonPropertyName("thirdPartyCollection")]
-    public bool? ThirdPartyCollection { get; set; }
-
-    /// <summary>
-    /// Total cancel delivery orders indicator
-    /// </summary>
-    [JsonPropertyName("totalCancelDeliveryOrders")]
-    public bool? TotalCancelDeliveryOrders { get; set; }
-
-    /// <summary>
-    /// Reverse delivery note indicator
-    /// </summary>
-    [JsonPropertyName("reverseDeliveryNote")]
-    public bool? ReverseDeliveryNote { get; set; }
-
-    /// <summary>
-    /// Reverse delivery note purpose code.
-    /// Valid values: 1-NOT OBLIGED TO ISSUE, 2-REFUSAL/CLERICAL ERROR,
-    /// 3-INTRA-COMMUNITY ACQUISITION, 4-THIRD COUNTRY ACQUISITION, 5-REVERSAL OF OBLIGATION
-    /// </summary>
-    [JsonPropertyName("reverseDeliveryNotePurpose")]
-    public int? ReverseDeliveryNotePurpose { get; set; }
-
-    /// <summary>
-    /// Other correlated entities
-    /// </summary>
-    [JsonPropertyName("otherCorrelatedEntities")]
-    public List<EntityOverride>? OtherCorrelatedEntities { get; set; }
-}
-
-public class OtherDeliveryNoteHeaderOverride
-{
-    /// <summary>
-    /// Loading address
-    /// </summary>
-    [JsonPropertyName("loadingAddress")]
-    public AddressOverride? LoadingAddress { get; set; }
-
-    /// <summary>
-    /// Delivery address
-    /// </summary>
-    [JsonPropertyName("deliveryAddress")]
-    public AddressOverride? DeliveryAddress { get; set; }
-
-    /// <summary>
-    /// Start shipping branch
-    /// </summary>
-    [JsonPropertyName("startShippingBranch")]
-    public int? StartShippingBranch { get; set; }
-
-    /// <summary>
-    /// Complete shipping branch
-    /// </summary>
-    [JsonPropertyName("completeShippingBranch")]
-    public int? CompleteShippingBranch { get; set; }
-}
-
-public class AddressOverride
-{
-    /// <summary>
-    /// Street name
-    /// </summary>
-    [JsonPropertyName("street")]
-    public string? Street { get; set; }
-
-    /// <summary>
-    /// Street number
-    /// </summary>
-    [JsonPropertyName("number")]
-    public string? Number { get; set; }
-
-    /// <summary>
-    /// Postal code
-    /// </summary>
-    [JsonPropertyName("postalCode")]
-    public string? PostalCode { get; set; }
-
-    /// <summary>
-    /// City name
-    /// </summary>
-    [JsonPropertyName("city")]
-    public string? City { get; set; }
-}
-
-/// <summary>
-/// Override for PartyType fields not mapped from cbCustomer or master data.
-/// Only fields without an existing general interface mapping are available.
-/// </summary>
-public class PartyUnmappedFieldsOverride
-{
-    [JsonPropertyName("documentIdNo")]
-    public string? DocumentIdNo { get; set; }
-
-    [JsonPropertyName("supplyAccountNo")]
-    public string? SupplyAccountNo { get; set; }
-
-    [JsonPropertyName("countryDocumentId")]
-    public string? CountryDocumentId { get; set; }
-}
-
-public class EntityOverride
-{
-    [JsonPropertyName("type")]
-    public int? Type { get; set; }
-
-    [JsonPropertyName("entityData")]
-    public PartyOverride? EntityData { get; set; }
-}
-
-public class PartyOverride
-{
-    [JsonPropertyName("vatNumber")]
-    public string? VatNumber { get; set; }
-
-    [JsonPropertyName("country")]
-    public string? Country { get; set; }
-
-    [JsonPropertyName("branch")]
-    public int? Branch { get; set; }
-
-    [JsonPropertyName("name")]
-    public string? Name { get; set; }
-
-    [JsonPropertyName("address")]
-    public AddressOverride? Address { get; set; }
-
-    [JsonPropertyName("documentIdNo")]
-    public string? DocumentIdNo { get; set; }
-
-    [JsonPropertyName("supplyAccountNo")]
-    public string? SupplyAccountNo { get; set; }
-
-    [JsonPropertyName("countryDocumentId")]
-    public string? CountryDocumentId { get; set; }
-}
-
-/// <summary>
-/// Line-level override for InvoiceRowType fields not set by the middleware.
-/// Applied per ChargeItem via ftChargeItemCaseData.
-/// </summary>
-public class InvoiceDetailOverride
-{
-    [JsonPropertyName("taricNo")]
-    public string? TaricNo { get; set; }
-
-    [JsonPropertyName("itemCode")]
-    public string? ItemCode { get; set; }
-
-    [JsonPropertyName("fuelCode")]
-    public int? FuelCode { get; set; }
-
-    [JsonPropertyName("lineComments")]
-    public string? LineComments { get; set; }
-
-    [JsonPropertyName("quantity15")]
-    public decimal? Quantity15 { get; set; }
-
-    [JsonPropertyName("otherMeasurementUnitQuantity")]
-    public int? OtherMeasurementUnitQuantity { get; set; }
-
-    [JsonPropertyName("otherMeasurementUnitTitle")]
-    public string? OtherMeasurementUnitTitle { get; set; }
-
-    [JsonPropertyName("notVAT195")]
-    public bool? NotVAT195 { get; set; }
-
-    [JsonPropertyName("dienergia")]
-    public ShipOverride? Dienergia { get; set; }
-
-    [JsonPropertyName("discountOption")]
-    public bool? DiscountOption { get; set; }
-}
-
-public class ShipOverride
-{
-    [JsonPropertyName("applicationId")]
-    public string? ApplicationId { get; set; }
-
-    [JsonPropertyName("applicationDate")]
-    public DateTime? ApplicationDate { get; set; }
-
-    [JsonPropertyName("doy")]
-    public string? Doy { get; set; }
-
-    [JsonPropertyName("shipId")]
-    public string? ShipId { get; set; }
-}
-
-/// <summary>
-/// Wrapper for line-level mydataoverride in ftChargeItemCaseData.
-/// JSON: { "GR": { "mydataoverride": { "invoiceDetails": { ... } } } }
-/// </summary>
-public class ChargeItemMyDataOverride
-{
-    [JsonPropertyName("invoiceDetails")]
-    public InvoiceDetailOverride? InvoiceDetails { get; set; }
-}
 
 public class MyDataSCU : IGRSSCD
 {
